export network related things to his own file.
